# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Goa Houseboat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all room packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a new package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a single package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageBody"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: addGalleryImage
      tags: [gallery]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryImageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostList"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /blog/{id}/approve:
    post:
      operationId: approveBlogPost
      tags: [blog]
      summary: Approve a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/slug/{slug}:
    get:
      operationId: getBlogPostBySlug
      tags: [blog]
      summary: Get blog post by slug (SEO-friendly URL)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blog/send-otp:
    post:
      operationId: sendBlogOtp
      tags: [blog]
      summary: Send OTP to email for anonymous blog submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /blog/verify-otp:
    post:
      operationId: verifyBlogOtp
      tags: [blog]
      summary: Verify OTP and get token for blog submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        pricePerNight:
          type: string
        capacity:
          type: integer
        inclusions:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - pricePerNight
        - capacity
        - inclusions
        - images
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
    CreatePackageBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pricePerNight:
          type: string
        capacity:
          type: integer
        inclusions:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
      required:
        - name
        - description
        - pricePerNight
        - capacity
    UpdatePackageBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pricePerNight:
          type: string
        capacity:
          type: integer
        inclusions:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
...
    GalleryImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: ["string", "null"]
        category:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - url
        - caption
        - category
        - sortOrder
        - createdAt
    CreateGalleryImageBody:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
      required:
        - url
        - category
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        authorName:
          type: string
        authorEmail:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, published, rejected]
        isAdminPost:
          type: boolean
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - content
        - excerpt
        - authorName
        - authorEmail
        - images
        - status
        - isAdminPost
        - metaTitle
        - metaDescription
        - publishedAt
        - createdAt
        - updatedAt
    BlogPostList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit
    CreateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        authorName:
          type: string
        authorEmail:
          type: string
        images:
          type: array
          items:
            type: string
        otpToken:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
      required:
        - title
        - content
        - authorName
    UpdateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, published, rejected]
        metaTitle:
          type: string
        metaDescription:
          type: string
    SendOtpBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    VerifyOtpBody:
      type: object
      properties:
        email:
          type: string
        otp:
          type: string
      required:
        - email
        - otp
    OtpVerifyResponse:
      type: object
      properties:
        token:
          type: string
        message:
          type: string
      required:
        - token
        - message
...