# lib/api-zod/src/generated/types/createPackageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Goa Houseboat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePackageBody {
  name: string;
  description: string;
  pricePerNight: string;
  capacity: number;
  inclusions?: string[];
  images?: string[];
  sortOrder?: number;
}